Update Jenkinsfile
Changes:
1. Fix link to github commit page in slack message.
2. Use list instead of variable for BUILD_IF_BRANCH (possible to configure several branches for build).
3. Remove skip ci plugin and use conditions instead. Allow avoid "ABORTED build" messages to slack.

+++ Jenkinsfile
@@ -1,5 +1,8 @@
-/* Change this variable to your baranch name if you want build it. */
-BUILD_IF_BRANCH = 'master'
+/*      Change this list with your baranch names if you want build it.
+        For example:
+def BUILD_IF_BRANCH = ['master','develop']  */
+
+def BUILD_IF_BRANCH = ['master']
 
 pipeline {
     agent { 
@@ -15,46 +18,55 @@ pipeline {
                 TEST_TARGET = "8.12.2"
             }
 
-   stages {
-        stage('Build') {
-	    when {
-		    branch "${BUILD_IF_BRANCH}"
-	    }
-            steps {
-
-
-		script {
-		    env.GIT_COMMIT_MSG = sh (script: 'git log -1 --pretty=%B ${GIT_COMMIT}', returnStdout: true).trim()
-		    env.GIT_AUTHOR = sh (script: 'git log -1 --pretty=%cn ${GIT_COMMIT}', returnStdout: true).trim()
-		    skip_ci_result = sh (script: "git log -1 | grep '.*\\[skip ci\\].*'", returnStatus: true)
-		    if (skip_ci_result == 0) {
-		        env.SKIP_CI = "true"
-		    }
+    stages {
+		stage('Prepare') {
+			steps {
+				script {
+					env.GIT_COMMIT_MSG = sh (script: 'git log -1 --pretty=%B ${GIT_COMMIT}', returnStdout: true).trim()
+					env.GIT_AUTHOR = sh (script: 'git log -1 --pretty=%cn ${GIT_COMMIT}', returnStdout: true).trim()
+					skip_ci_result = sh (script: "git log -1 | grep '.*\\[skip ci\\].*'", returnStatus: true)
+					if (skip_ci_result == 0) {
+						env.SKIP_CI = "true"
+						echo "Build going to be skiped due to [skip ci] tag"
+					}
+					if (env.BRANCH_NAME in BUILD_IF_BRANCH) {
+						env.SKIP_BRANCH = "false"
+					} else {
+						env.SKIP_BRANCH = "true"
+						echo "Build going to be skiped due to branche name"
+					  }
+				}
+			}
 		}
 
-		scmSkip(deleteBuild: true, skipPattern:'.*\\[skip ci\\].*')
-
-                sh '''
-                      eval $(opam config env)
-                      opam config var root
-                      if [ -d lib/vellvm ]; then rm -rf lib/vellvm; fi
-                      git clone --recurse-submodules --shallow-submodules --depth 1 https://github.com/vellvm/vellvm.git lib/vellvm
-                      git --no-pager --git-dir=lib/vellvm/.git log -1 --pretty=oneline
-                      make -j ${NJOBS} -C lib/vellvm/src
-                      ln -s `pwd`/lib/vellvm/src/ml/libvellvm/ ml/
-                      make -j ${NJOBS}
-                      make test
-                   '''
-            }
-        }
+        stage('Build') {
+			steps {
+				script {
+					if (env.SKIP_CI != "true" && env.SKIP_BRANCH == "false") {
+						sh '''
+						    eval $(opam config env)
+  						    opam config var root
+						    if [ -d lib/vellvm ]; then rm -rf lib/vellvm; fi
+						    git clone --recurse-submodules --shallow-submodules --depth 1 https://github.com/vellvm/vellvm.git lib/vellvm
+						    git --no-pager --git-dir=lib/vellvm/.git log -1 --pretty=oneline
+						    make -j ${NJOBS} -C lib/vellvm/src
+						    ln -s `pwd`/lib/vellvm/src/ml/libvellvm/ ml/
+						    make -j ${NJOBS}
+						    make test
+						  '''
+					}
+				}
+			}
+		}
     }
-    post {
-       always {
-	   /* Use slackNotifier.groovy from shared library and provide current build result as parameter */
+
+	post {
+        always {
+	       /* Use slackNotifier.groovy from shared library and provide current build result as parameter */
            notifySlack(currentBuild.result)
            cleanWs()
-       }
-   }
+        }
+    }
 }
 
 /*   Functions   */
@@ -78,25 +90,26 @@ def notifySlack(String buildStatus = 'STARTED') {
 
     def msg = "Build <${env.BUILD_URL}|#${env.BUILD_NUMBER}> (${getSlackRepoURL()}) of ${env.JOB_NAME} (${env.GIT_BRANCH}) by ${env.GIT_AUTHOR} ${buildStatus} in ${currentBuild.durationString.minus(' and counting')}"
 
-	if (env.BRANCH_NAME == "${BUILD_IF_BRANCH}" && !(env.SKIP_CI == "true")) {
-	  slackSend(color: color, message: msg, channel: '#bitbucket-activity')
+	if (env.SKIP_BRANCH == "false" && !(env.SKIP_CI == "true")) {
+		slackSend(color: color, message: msg, channel: '#bitbucket-activity')
     }
 }
 
 def getRepoURL() {
-  sh "git config --get remote.origin.url > .git/remote-url"
-  return readFile(".git/remote-url").trim()
+	sh "git config --get remote.origin.url > .git/remote-url"
+	return readFile(".git/remote-url").trim()
 }
 
 def getRepoSHA() {
-  sh "git rev-parse HEAD > .git/current-commit"
-  return readFile(".git/current-commit").trim()
+	sh "git rev-parse HEAD > .git/current-commit"
+	return readFile(".git/current-commit").trim()
 }
 
 def getSlackRepoURL() {
-  repoURL = getRepoURL()
-  repoURL = repoURL.take(repoURL.length()-4) + "/commit"
-  repoSHA = getRepoSHA()
-  
-	return "<${repoURL}|${getRepoSHA().take(7)}>"
+	repoURL = getRepoURL()
+	repoURL = repoURL.take(repoURL.length()-4) + "/commit"
+	repoSHA = getRepoSHA()
+
+	return "<${repoURL}/${repoSHA}|${getRepoSHA().take(7)}>"
 }
+
